added extra layer in cs_google: compute_engine and moved authentication in actions/authentication
